chore: run "sync" workflow only on main branch (GEA-11243)
The README says that the "sync" workflow only works on the configured
GitHub default branch but then the project itself doesn't adhere to
this. This causes all PR branches to have a failed "sync" run. This
patch fixes this by updating the workflow to only run on the default
branch.